Harden the authorization server on .NET 11
- Move to .NET 11 with Lumoin.Base and Lumoin.Base.Libsodium 0.0.14;
  libsodium runs on browser-wasm behind a pinned scratch pool, proven
  by a smoke under the workload's Node.
- Accept RFC 8252 loopback redirects on any port behind a canonical
  form gate; localhost by policy opt-in only.
- Redeem an authorization code exactly once; rotate refresh tokens
  exactly once; a reused rotated-out token revokes the grant family
  through an uncapped, claimed walk; one client-identification rule;
  constant refusal bodies; a jti-keyed revocation seam beside the
  RFC 7009 one; the delete and endpoint-URI seams required.
- The JSONata engine behind a byte-typed seam; the first-party engine
  and the JSON Pointer package removed.
- Sweep the style rules across the tree at the 10.0 analysis level.
- The sixteen tests that alter a running host gathered under one
  name, with the live-configuration design in the design document.

+++ src/Verifiable.Cbor/Mdoc/MdocCborDeviceSignedSigner.cs
@@ -1,5 +1,5 @@
-using System.Buffers;
 using Lumoin.Veritas.Cbor;
+using System.Buffers;
 using Verifiable.Core.Model.Mdoc;
 using Verifiable.Cryptography;
 using Verifiable.JCose;
@@ -178,7 +178,7 @@ public static class MdocCborDeviceSignedSigner
         using var buffer = new SlabBufferWriter(pool);
         var writer = new CborWriter(buffer, CborOptions.RfcCanonical);
 
-        writer.WriteTag(new CborTag((ulong)CoseTags.Sign1));
+        writer.WriteTag(new CborTag(CoseTags.Sign1));
         writer.WriteStartArray(4);
         writer.WriteByteString(message.ProtectedHeader.AsReadOnlySpan());
         writer.WriteStartMap(0);
